Copy in theme HTML and CSS files
This is the latest state of the work-in-progress Operations guide, which is where this theme was developed. The plan is to move that project to use this Gem, and future enhancements to layout and new features will be introduced at the Gem level, rather than at the site level (unless it's a very site-specific thing, of course).

More abstraction will be needed before this is ready for use outside of ACCD, or even outside of the Operations Guides, but this lays the foundation for those efforts.