Add PDF layout
This layout file and corresponding CSS compiles a site's homepage content and all sections and sub-sections into a single HTML page, accessible at `/pdf` off the baseurl (for example, https://ops.palmereventscenter.com/pdf), which inherits print-optimized styles. That page can be provided to an HTML-to-PDF converter (either desktop, command-line, or web-based) to generate a single PDF of the entire site.

**Required for maintainers:**
You must add `pdf.md` to the root of your project (at the same level as `index.md`). Do not place any content or frontmatter in the file.