Add config for hiding pages from search results
In our use case of the Kocurek theme there are some sites we want indexed by Google (like the Operational Policies), and there are some sites we might prefer not get crawled (like the Content Management Guides).

Adding `hidden_from_search: true` to a site's `_config.yml` will add the appropriate `<meta>` tag to each page to tell search engines to treat the page as though it does not exist.